requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extensions==3.1.1
django-organizations==2.0.0
pkg-resources==0.0.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
